AbortController가 적용되는 fetchBody 메서드 구현(aboratlbeFetchBody)
