+++ src/model/label/Api.js
@@ -1,34 +1,8 @@
-import { fetchBody } from "../../utils";
+import { fetchBody, abortableFetchBody } from "../../utils";
 
 const getLabels = () => fetchBody("/labels");
 
-const throwIfNotAbortError = error => {
-  if(error.name !== 'AbortError') {
-    throw error;
-  }
-}
-
-const getLabelsDelay = (() => {
-  let requested = false;
-  let controller;
-  
-  return () => {
-    if(requested) {
-      controller.abort();
-    }
-    controller = new AbortController();
-    const {signal} = controller;
-    requested = true;
-
-
-    return fetchBody("/labels-delay", {signal})
-      .then((res) => {
-        requested = false
-        return res;
-      })
-      .catch(throwIfNotAbortError)
-  }
-})();
+const getLabelsDelay = () => abortableFetchBody('/labels-delay');
 
 const addLabels = (label) =>
   fetchBody("/labels", {
