+++ src/utils/index.js
@@ -23,6 +23,32 @@ const removeClassList = ({
 element.querySelector(selector).classList.remove(classes);
 
 const fetchBody = (url, options={}) => fetch(url, options).then((response) => response.json());
+
+const throwIfNotAbortError = error => {
+  if(error.name !== 'AbortError') {
+    throw error;
+  }
+}
+
+/** 요청이 완료되기전에 재시도하면 기존 요청은 취소하고 다시 요청하는 fetchBody */
+const abortableFetchBody = (() => {
+  
+  let controller;
+  let requested = false;
+  
+  return (url, options = {}) => {
+    if(requested) {
+      controller.abort();
+    }
+    controller = new AbortController();
+    const {signal} = controller;
+    const _options = {...options, signal};
+    requested = true;
+    fetchBody(url, _options).catch(throwIfNotAbortError)
+    .finally(() => requested = false);
+  }
+})()
+
 const pipe = (...fns) => {
   return (args) =>
     fns.reduce((acc, fn) => {
@@ -84,6 +110,7 @@ export {
   addClassList,
   removeClassList,
   fetchBody,
+  abortableFetchBody,
   pipe,
   cacheFunction,
   getRandomColor,
